requirements.txt
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
tradingview-screener
prophet
yfinance
matplotlib
python-dotenv
hyperliquid-python-sdk
eth-account
psycopg2-binary
openai>=1.0.0
apscheduler>=3.10.0
requests>=2.28.0
